fixing issue #117 (thanks Anton!)
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@657 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ surface/include/pcl/surface/impl/concave_hull.hpp
@@ -248,7 +248,7 @@ pcl::ConcaveHull<PointInT>::performReconstruction (PointCloud &alpha_shape,
   int max_vertex_id=-1;
   FORALLvertices
   {
-    if (vertex->id > max_vertex_id)
+    if ((int)vertex->id > max_vertex_id)
       max_vertex_id = vertex->id;
   }
